+++ lib/audio_metadata_reader.dart
@@ -4,4 +4,3 @@ export "src/parsers/tag_parser.dart" show AudioMetadata, Picture;
 export 'src/metadata/base.dart' show PictureType;
 export 'src/utils/metadata_parser_exception.dart' show MetadataParserException;
 export 'src/parser.dart' show readMetadata;
-export 'src/writer.dart' show writeMetadata;