+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 ## 0.0.7
 
-- BREAKING CHANGE : the API is now synchronous
+### BREAKING CHANGE
+
+- `readMetadata` is now synchronous. You may remove some `Future.wait()`
+- `InvalidTag` class has beem removed and replace with a real exception `MetadataParserException`
+
+## Other
+
 - Performance improvement : 70% faster by using sync IO operations and a File Buffer to read files
 
 ## 0.0.6